upwork-automation/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
setuptools
aiohttp
requests
pandas
numpy
lxml  
tiktoken 

# IBM watsonx Dependencies (HACKATHON CRITICAL)
# Note: Python 3.11 - Full support for latest IBM packages
ibm-watsonx-ai>=1.0.0
ibm-watson-machine-learning>=1.0.350
ibm-generative-ai>=2.0.0

# IBM watsonx Orchestrate ADK (HACKATHON CRITICAL)
# Note: Install from IBM's repository
# pip install ibm-watsonx-orchestrate-adk

# Vector Store (Keep FAISS as backup)
faiss-cpu 
langchain-text-splitters
langchain-ibm>=0.1.0

# Keep for comparison/fallback (OpenAI used in rag_relevance.py)
openai
langchain-openai
